Rewrite Enterprise PageCache to stop container being saved with JS in.
The observer event for saving container to cache happens before our observer.  There's no way of changing this dependency
so that means a rewrite is required.

It also means that in the instances that we are dealing with html from a container we have to assume that the current process will move HTML
to the end of the document correctly and then be saved in the FPC.  Which means we have to remove the JS and then save that to the container cache.

No so pretty, but appears to be functioning.  The main piece of JS that causes issue is the minicar truncateOptions().  This code does have a formkey in it
but as this doesn't changed for the customer after being initially generated the functionality is maintained.